extract TextAreaVerticalScrollbar
